Add LLVM fuzzer harness
